numberwiki/requirements.txt
Flask==2.0.3
flask-restful==0.3.9
flask_api
pyyaml
requests